requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
Pillow>=10.0.0
plotly>=5.17.0
numpy>=1.24.0
python-dateutil>=2.8.2
instaloader>=4.10.0  